pyproject.toml
[project]
name = "oxonfair"
version = "0.2.1.9"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.8",
    "fairlearn>=0.12.0",
    "lightgbm>=4.6.0",
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
    "xgboost>=3.0.0",
]

[dependency-groups]
vision = [
    "h5py>=3.13.0",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pytorch-lightning>=2.5.1",
    "tabulate>=0.9.0",
    "torchvision>=0.21.0",
]
